==> notebooks/01_load_and_clean.ipynb <==
import pandas as pd

df_val_lq = pd.read_excel("../data/raw/excel/Company_Screening_Report_Valuation_LQ.xlsx")
df_val_lq.to_csv("../data/raw/fundamentals_tev_LQ.csv", index=False)

df_val_historical = pd.read_excel("../data/raw/excel/Company_Screening_Report_Valuation_Historical.xlsx")
df_val_historical.to_csv("../data/raw/fundamentals_tev_historical.csv", index=False)

df_CY2018 = pd.read_excel("../data/raw/excel/Company_Screening_Report_Financials_CY2018.xlsx")
df_CY2018.to_csv("../data/raw/fundamentals_CY2018.csv", index=False)

df_CY2019 = pd.read_excel("../data/raw/excel/Company_Screening_Report_Financials_CY2019.xlsx")
df_CY2019.to_csv("../data/raw/fundamentals_CY2019.csv", index=False)

df_CY2020 = pd.read_excel("../data/raw/excel/Company_Screening_Report_Financials_CY2020.xlsx")
df_CY2020.to_csv("../data/raw/fundamentals_CY2020.csv", index=False)

df_CY2021 = pd.read_excel("../data/raw/excel/Company_Screening_Report_Financials_CY2021.xlsx")
df_CY2021.to_csv("../data/raw/fundamentals_CY2021.csv", index=False)

df_CY2022 = pd.read_excel("../data/raw/excel/Company_Screening_Report_Financials_CY2022.xlsx")
df_CY2022.to_csv("../data/raw/fundamentals_CY2022.csv", index=False)

df_CY2023 = pd.read_excel("../data/raw/excel/Company_Screening_Report_Financials_CY2023.xlsx")
df_CY2023.to_csv("../data/raw/fundamentals_CY2023.csv", index=False)

df_CY2024 = pd.read_excel("../data/raw/excel/Company_Screening_Report_Financials_CY2024.xlsx")
df_CY2024.to_csv("../data/raw/fundamentals_CY2024.csv", index=False)

df_LQ = pd.read_excel("../data/raw/excel/Company_Screening_Report_Financials_LQ.xlsx")
df_LQ.to_csv("../data/raw/fundamentals_LQ.csv", index=False)

df_val_lq.head(10)

df_CY2018.head(10)

def clean(df, time_period):
    df.columns = df.iloc[6]
    df.columns.name = None
    df.drop(df.index[0:7],inplace=True)
    df.reset_index(drop=True,inplace=True)
    df = df.rename(columns={"Security Tickers":"Time Period"})
    df["Time Period"] = time_period
    df["Company Name"] = df["Company Name"].str.split("(").str[0]
    df["Company Name"] = df["Company Name"].str[:-1]
    df.columns = (df.columns.str.replace(r"\s*\[.*?\]\s*", " ", regex=True).str.strip())
    df[["Exchange1", "Ticker1"]] = df["Exchange:Ticker"].str.split(":", n=1, expand=True)
    df.insert(1, "Ticker", df["Ticker1"])
    df.insert(1, "Exchange", df["Exchange1"])
    df = df.drop(columns=["Exchange:Ticker", "Exchange1", "Ticker1"])
    return df

df_val_lq = clean(df_val_lq, "Latest Quarter")
df_val_historical = clean(df_val_historical, "N/A")
df_CY2018 = clean(df_CY2018, "2018")
df_CY2019 = clean(df_CY2019, "2019")
df_CY2020 = clean(df_CY2020, "2020")
df_CY2021 = clean(df_CY2021, "2021")
df_CY2022 = clean(df_CY2022, "2022")
df_CY2023 = clean(df_CY2023, "2023")
df_CY2024 = clean(df_CY2024, "2024")
df_LQ = clean(df_LQ, "Latest Quarter")

df_LQ.columns

df_val_lq.head(10)

df_val_lq.to_csv("../data/processed/fundamentals_clean_val_lq_only.csv", index=False)

df_val_historical.columns

cols = df_val_historical.columns.tolist()
cols[7:] = [col[:23] + col[24:] for col in cols[7:]]
df_val_historical.columns = cols

df_val_historical.columns

df_val_historical.insert(loc=len(df_val_historical.columns), column="Total Enterprise Value ($USDmm, Latest Quarter)", value=df_val_lq["Total Enterprise Value  ($USDmm, Historical rate)"])

df_val_historical.head(10)

df_val_historical.columns

df_CY2018.loc[0:5,["Company Name"]]

df_CY2018.head(10)

df_val_historical.to_csv("../data/processed/fundamentals_clean_val.csv", index=False)
df_CY2018.to_csv("../data/processed/fundamentals_clean_CY2018.csv", index=False)
df_CY2019.to_csv("../data/processed/fundamentals_clean_CY2019.csv", index=False)
df_CY2020.to_csv("../data/processed/fundamentals_clean_CY2020.csv", index=False)
df_CY2021.to_csv("../data/processed/fundamentals_clean_CY2021.csv", index=False)
df_CY2022.to_csv("../data/processed/fundamentals_clean_CY2022.csv", index=False)
df_CY2023.to_csv("../data/processed/fundamentals_clean_CY2023.csv", index=False)
df_CY2024.to_csv("../data/processed/fundamentals_clean_CY2024.csv", index=False)
df_LQ.to_csv("../data/processed/fundamentals_clean_LQ.csv", index=False)
