==> notebooks/02_build_panel.ipynb <==
import pandas as pd
import numpy as np
import sys
sys.path.append("../src")
from yoy_data_builder import *

df_CY2018 = pd.read_csv("../data/processed/fundamentals_clean_CY2018.csv")
df_CY2019 = pd.read_csv("../data/processed/fundamentals_clean_CY2019.csv")
df_CY2020 = pd.read_csv("../data/processed/fundamentals_clean_CY2020.csv")
df_CY2021 = pd.read_csv("../data/processed/fundamentals_clean_CY2021.csv")
df_CY2022 = pd.read_csv("../data/processed/fundamentals_clean_CY2022.csv")
df_CY2023 = pd.read_csv("../data/processed/fundamentals_clean_CY2023.csv")
df_CY2024 = pd.read_csv("../data/processed/fundamentals_clean_CY2024.csv")
df_LQ = pd.read_csv("../data/processed/fundamentals_clean_LQ.csv")

dfs = [df_CY2018, df_CY2019, df_CY2020, df_CY2021, df_CY2022, df_CY2023, df_CY2024, df_LQ]
dfs = [df.assign(row_index=df.index) for df in dfs]
dfs = [df.assign(dataset_id=i) for i, df in enumerate(dfs)]
df = pd.concat(dfs).sort_values(by=["row_index", "dataset_id"]).reset_index(drop=True)
df = df.drop(columns=["row_index", "dataset_id"])
df = df.rename(columns={"Total Revenue ($USDmm, Historical rate)": "Total Revenue ($USDmm)"})
df = df.rename(columns={"Capital Expenditure ($USDmm, Historical rate)": "Capital Expenditure ($USDmm)"})
df["Total Revenue ($USDmm)"] = pd.to_numeric(df["Total Revenue ($USDmm)"], errors="coerce")
df["Gross Margin %"] = pd.to_numeric(df["Gross Margin %"], errors="coerce")
df["Capital Expenditure ($USDmm)"] = pd.to_numeric(df["Capital Expenditure ($USDmm)"], errors="coerce")

df.head(24)

df.to_csv("../data/processed/combined_financials_with_LQ.csv", index=False)

#drop LQ for historical fundamentals
df = df[df["Time Period"] != "Latest Quarter"]
df = df.rename(columns={"Time Period": "Year"})

revenue_yoy(df)
capex_yoy(df)

df.head(10)

df["Capex to Rev"] = (df["Capital Expenditure ($USDmm)"]/df["Total Revenue ($USDmm)"]).round(4)

df.head(10)

df.to_csv("../data/processed/fundamentals_panel.csv", index=False)
