==> notebooks/03_stage_definition.ipynb <==
import pandas as pd
import numpy as np
%load_ext autoreload
%autoreload 2
import sys
sys.path.append("../src")
from stage_classifier import stage_classifier

df_fundamentals = pd.read_csv("../data/processed/fundamentals_panel.csv")

df_fundamentals.head()

df_fundamentals.dtypes

# Break into stages based on Revenue YoY, Capital Intensity, Gross Margin %, and Revenue
df_fundamentals = stage_classifier(df_fundamentals)

df_fundamentals.head(20)

df_fundamentals.to_csv("../data/processed/fundamentals_panel_grouped.csv")

df_tev = pd.read_csv("../data/processed/fundamentals_clean_val.csv")

df_tev.head(10)

df_tev.dtypes

df_tev["Total Enterprise Value ($USDmm, 2018)"] = pd.to_numeric(df_tev["Total Enterprise Value ($USDmm, 2018)"], errors="coerce")
df_tev["Total Enterprise Value ($USDmm, 2019)"] = pd.to_numeric(df_tev["Total Enterprise Value ($USDmm, 2019)"], errors="coerce")
df_tev["Total Enterprise Value ($USDmm, 2020)"] = pd.to_numeric(df_tev["Total Enterprise Value ($USDmm, 2020)"], errors="coerce")
df_tev["Total Enterprise Value ($USDmm, 2021)"] = pd.to_numeric(df_tev["Total Enterprise Value ($USDmm, 2021)"], errors="coerce")
df_tev["Total Enterprise Value ($USDmm, 2022)"] = pd.to_numeric(df_tev["Total Enterprise Value ($USDmm, 2022)"], errors="coerce")
df_tev["Total Enterprise Value ($USDmm, 2023)"] = pd.to_numeric(df_tev["Total Enterprise Value ($USDmm, 2023)"], errors="coerce")
df_tev["Total Enterprise Value ($USDmm, 2024)"] = pd.to_numeric(df_tev["Total Enterprise Value ($USDmm, 2024)"], errors="coerce")
df_tev["Total Enterprise Value ($USDmm, Latest Quarter)"] = pd.to_numeric(df_tev["Total Enterprise Value ($USDmm, Latest Quarter)"], errors="coerce")
df_tev.dtypes

df_tev_merge = df_tev.copy()
df_tev_merge.drop(columns="Total Enterprise Value ($USDmm, Latest Quarter)",inplace=True)
cols = df_tev_merge.columns.tolist()
cols[7:] = [col[-5:-1] for col in cols[7:]]
df_tev_merge.columns = cols
df_tev_merge = df_tev_merge.iloc[:,[2,7,8,9,10,11,12,13]]

df_tev_merge.head()

df_tev_merge = df_tev_merge.melt(id_vars=["Ticker"], var_name="Year", value_name="TEV")
df_tev_merge["Year"] = df_tev_merge["Year"].astype("int64")

df_tev_merge.head(24)

df_fundamentals = df_fundamentals.merge(df_tev_merge,on=["Ticker","Year"],how="left")

df_fundamentals.head(10)

df_fundamentals["EV/Rev"] = df_fundamentals["TEV"] / df_fundamentals["Total Revenue ($USDmm)"]

df_fundamentals.head(24)

df_fundamentals = df_fundamentals[
    [
        "Company Name",
        "Exchange",
        "Ticker",
        "Year",
        "CIK",
        "Excel Company ID",
        "Industry Classifications",
        "Total Revenue ($USDmm)",
        "Revenue YoY",
        "Capex YoY",
        "Gross Margin %",
        "Capital Expenditure ($USDmm)",
        "Capex to Rev",
        "Growth Bucket",
        "Capex Bucket",
        "Margin Bucket",
        "Revenue Bucket",
        "TEV",
        "EV/Rev"
    ]
]
df_fundamentals.head()

df_fundamentals.to_csv("../data/processed/full_panel.csv")
