==> notebooks/04_sql_table_outputs.ipynb <==
import sqlite3
import pandas as pd
from pathlib import Path
import numpy as np
%load_ext autoreload
%autoreload 2

import sys
sys.path.append("../src")
from call_sql_queries import *

df_cbg = coverage_by_growth(None)
df_cbg.to_csv("../outputs/notebook_outputs/coverage_by_growth.csv")
df_cbg

df_cbm = coverage_by_margin(None)
df_cbm.to_csv("../outputs/notebook_outputs/coverage_by_margin.csv")
df_cbm

df_cbreg = coverage_by_regime(None, None, None, None, None, None)
df_cbreg.to_csv("../outputs/notebook_outputs/coverage_by_regime_all.csv")
df_cbreg

df_cbrmc = coverage_by_regime_minus_capex(None, None, None, None, None)
df_cbrmc.to_csv("../outputs/notebook_outputs/coverage_by_regime_no_capex.csv")
df_cbrmc

coverage_by_regime("High Growth", "Asset Light", "Low Margin", "$100-250M", None, None)

df_vbr = valuation_by_regime(None, None, None, None, None, None)
df_vbr.to_csv("../outputs/notebook_outputs/valuation_by_regime_all.csv")
df_vbr

df_vbgc = valuation_by_growth_capex()
df_vbgc.to_csv("../outputs/notebook_outputs/valuation_by_growth_capex.csv")
df_vbgc

df_vbgm = valuation_by_growth_margin()
df_vbgm.to_csv("../outputs/notebook_outputs/valuation_by_growth_margin.csv")
df_vbgm

df_vbgr = valuation_by_growth_revenue()
df_vbgr.to_csv("../outputs/notebook_outputs/valuation_by_growth_revenue.csv")
df_vbgr

df_cavbcs = coverage_and_valuation_by_current_stage(None, None)
df_cavbcs.to_csv("../outputs/notebook_outputs/coverage_and_valuation_by_current_stage_all.csv")
df_cavbcs

# testing below

pull_comps_exact("Low Growth", "Moderate Capital Intensity", "Very High Margin", "$50-100M")

pull_comps_fuzzy("Low Growth", "Moderate Capital Intensity", "Very Low Margin", "$50-100M")

pull_stage_comps("Low Growth", "$50-100M")

pull_stage_comps_coverage_median_mean("Low Growth", "$50-100M")

pull_stage_comps_plus_margin_coverage_and_median("Low Growth", "Very High Margin", "$50-100M")

pull_stage_comps_coverage_and_median_by_margin_bracket("Low Growth", "$50-100M")
