==> notebooks/05_company_class_test.ipynb <==
import pandas as pd
import sys
sys.path.append("../src")
from company_blueprint import Company
%load_ext autoreload
%autoreload 2

df_apld = Company("Applied Digital", "NASDAQGS", "APLD", 173.6, .284, .226, -686.6, 8718.2)

# get_company_data()

# exact_comps_coverage()

# exact_comps()

# fuzzy_comps()
       
# stage_comps()
        
# exact_comps_valuation()
       
# stage_comps_coverage_and_valuation()
  
# stage_comps_coverage_median_mean()
     
# stage_comps_plus_margin_coverage_and_median()
    
# stage_comps_coverage_and_median_by_margin_bracket()
        
# get_ev_to_rev()

# comps_minus_capex_coverage()

# comps_minus_capex()

# fuzzy_comps_minus_capex()

df_apld.get_company_data()

df_apld.exact_comps_coverage()

df_apld.exact_comps()

df_apld.fuzzy_comps()

df_apld.stage_comps()

df_apld.exact_comps_valuation()

df_apld.stage_comps_coverage_and_valuation()

df_apld.stage_comps_coverage_median_mean()

df_apld.stage_comps_plus_margin_coverage_and_median()

df_apld.stage_comps_coverage_and_median_by_margin_bracket()

df_apld.comps_minus_capex_coverage()

df_apld.comps_minus_capex()

df_apld.fuzzy_comps_minus_capex()

df_apld.change_in_ev_rev_by_margin()

df_apld.valuation_growth_vs_narrative()
